requirements.txt
asgiref==3.3.4
click==8.0.1
dataclasses==0.8
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
Pillow==8.2.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
